Improve file direction
