+++ AdvancedSharpAdbClient/DeviceCommands/Models/AndroidProcess.cs
