+++ AdvancedSharpAdbClient/DeviceCommands/Models/AndroidProcessState.cs
