+++ AdvancedSharpAdbClient/DeviceCommands/Models/VersionInfo.cs
