+++ AdvancedSharpAdbClient/DeviceCommands/Receivers/EnvironmentVariablesReceiver.cs
