+++ AdvancedSharpAdbClient/DeviceCommands/Receivers/GetPropReceiver.cs
