+++ AdvancedSharpAdbClient/DeviceCommands/Receivers/InfoReceiver.cs
