+++ AdvancedSharpAdbClient/DeviceCommands/Receivers/InstallReceiver.cs
