+++ AdvancedSharpAdbClient/DeviceCommands/Receivers/PackageManagerReceiver.cs
