+++ AdvancedSharpAdbClient/DeviceCommands/Receivers/VersionInfoReceiver.cs
