+++ AdvancedSharpAdbClient/Extensions/Factories.cs
