+++ AdvancedSharpAdbClient/Extensions/NamespaceDoc.cs
