+++ AdvancedSharpAdbClient/Extensions/SyncCommandConverter.cs
