+++ AdvancedSharpAdbClient/Interfaces/IAdbCommandLineClient.cs
