+++ AdvancedSharpAdbClient/Interfaces/IAdbServer.cs
