+++ AdvancedSharpAdbClient/Interfaces/IAdbSocket.cs
