+++ AdvancedSharpAdbClient/Interfaces/IDeviceMonitor.cs
