+++ AdvancedSharpAdbClient/Interfaces/ISyncService.cs
