+++ AdvancedSharpAdbClient/Interfaces/ITcpSocket.cs
