+++ AdvancedSharpAdbClient/Models/AdbResponse.cs
