+++ AdvancedSharpAdbClient/Models/AdbServerFeatures.cs
