+++ AdvancedSharpAdbClient/Models/AdbServerStatus.cs
