+++ AdvancedSharpAdbClient/Models/ColorData.cs
