+++ AdvancedSharpAdbClient/Models/DeviceData.cs
