+++ AdvancedSharpAdbClient/Models/Enums/DeviceState.cs
