+++ AdvancedSharpAdbClient/Models/Enums/ForwardProtocol.cs
