+++ AdvancedSharpAdbClient/Models/Enums/StartServerResult.cs
