+++ AdvancedSharpAdbClient/Models/Enums/SyncCommand.cs
