+++ AdvancedSharpAdbClient/Models/Enums/TransportType.cs
