+++ AdvancedSharpAdbClient/Models/Enums/UnixFileMode.cs
