+++ AdvancedSharpAdbClient/Models/FileStatistics.cs
