+++ AdvancedSharpAdbClient/Models/ForwardData.cs
