+++ AdvancedSharpAdbClient/Models/ForwardSpec.cs
