+++ AdvancedSharpAdbClient/Models/FramebufferHeader.cs
